+++ app/system/internal/handler/invest/list/update_handler.go
@@ -25,7 +25,7 @@ func UpdateHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
 		if err != nil {
 			httpx.ErrorCtx(r.Context(), w, err)
 		} else {
-			httpx.Ok(w)
+			httpx.OkJsonCtx(r.Context(), w, nil)
 		}
 	}
 }