+++ app/system/internal/handler/product/config/info_handler.go
@@ -1,31 +1,31 @@
 // Code scaffolded by goctl. Safe to edit.
 // goctl 1.10.0
 
-package list
+package config
 
 import (
 	"net/http"
 
 	"github.com/zeromicro/go-zero/rest/httpx"
-	"ovra/app/system/internal/logic/invest/list"
+	"ovra/app/system/internal/logic/product/config"
 	"ovra/app/system/internal/svc"
 	"ovra/app/system/internal/types"
 )
 
-func DeleteHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
+func InfoHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		var req types.IdsReq
+		var req types.IdReq
 		if err := httpx.Parse(r, &req); err != nil {
 			httpx.ErrorCtx(r.Context(), w, err)
 			return
 		}
 
-		l := list.NewDeleteLogic(r.Context(), svcCtx)
-		err := l.Delete(&req)
+		l := config.NewInfoLogic(r.Context(), svcCtx)
+		resp, err := l.Info(&req)
 		if err != nil {
 			httpx.ErrorCtx(r.Context(), w, err)
 		} else {
-			httpx.OkJsonCtx(r.Context(), w, nil)
+			httpx.OkJsonCtx(r.Context(), w, resp)
 		}
 	}
 }